more

+++ sgl-kernel/csrc/gemm/per_token_group_quant_8bit.cu
@@ -139,10 +139,10 @@ void sgl_per_token_group_quant_8bit(
     dim3 block(num_threads);                                                                      \
     if (is_column_major) {                                                                        \
       if (scale_ue8m0) {                                                                          \
-        per_token_group_quant_8bit_kernel<T, float, DST_DTYPE, true, true><<<grid, block, 0, stream>>>(  \
+        per_token_group_quant_8bit_kernel<T, uint8_t, DST_DTYPE, true, true><<<grid, block, 0, stream>>>(  \
             static_cast<T*>(input.data_ptr()),                                                    \
             output_q.data_ptr(),                                                                  \
-            static_cast<float*>(output_s.data_ptr()),                                             \
+            static_cast<uint8_t*>(output_s.data_ptr()),                                             \
             group_size,                                                                           \
             num_groups,                                                                           \
             groups_per_block,                                                                     \
